Update frontend.py
Changed function stream_title_changed to pass. This function is intended to update the title variable when it sees a change, but it pulls stale data (at least when using the Jellyfin backend), which causes the displayed title to revert back to the previous track's title. This is a redundant function anyway due to the display update loop, so I have simply removed the variable update line and set the function to simply pass.